Time: 83 ms (19.79%), Space: 56.3 MB (44.33%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.java
@@ -1,7 +1,7 @@
 class Solution {
     public long repairCars(int[] ranks, int cars) {
         Arrays.sort(ranks);
-        long minimumTime = 1, maximumTime = ranks[ranks.length - 1] * (long)cars * (long)cars;
+        long minimumTime = 1, maximumTime = (long)ranks[ranks.length - 1] * cars * cars;
         while(minimumTime < maximumTime){
             long mid = (minimumTime + maximumTime) / 2;
             long sum = 0;
